Correlate in float32 in bm_covary (#5)
Removes two of the three float64 FFTs from the HASQI path. bm_covary reduces
its correlation to a magnitude, so the error degrades smoothly: worst relative
error over the golden set goes from 2.81e-08 to 1.42e-07.

That worst case is duration_0.05s, where 1200 samples leave too few segments to
average the error down; it gets its own tolerance rather than loosening the bar
for the other 34. Aligns seeded_parity_test.py with the 1e-7 the rest of the
suite already uses.

envelope_align still requires float64 and is unchanged.

+++ src/jax_haspi_hasqi/hasqi.py
@@ -219,7 +219,9 @@ def bm_covary(
     processed_ms = jnp.sum(processed_seg**2, axis=-1) * power
 
     length = reference_seg.shape[-1]
-    correlation = filters.correlate_full(reference_seg, processed_seg)
+    correlation = filters.correlate_full(
+      reference_seg, processed_seg, dtype=jnp.float32
+    )
     correlation = correlation[length - 1 - max_lag : max_lag + length]
     peak = jnp.max(jnp.abs(correlation * correction))
